cocoindex/scripts/pyproject.toml
[project]
name = "mdb-cocoindex"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "cocoindex",
    "openai>=2.16.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv",
    "voyageai>=0.3.7",
]
